Create Parser, FractalGenerator
