+++ NNPTPZ1/NNPTPZ1.csproj
@@ -45,6 +45,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ComplexNumber.cs" />
+    <Compile Include="FractalGenerator.cs" />
+    <Compile Include="Parser.cs" />
     <Compile Include="Polynomial.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
